Moved exceptions to own namespace
