+++ src/ContaoCommunityAlliance/Composer/Plugin/Exception/ConfigUpdateException.php
@@ -13,12 +13,11 @@
  * @license LGPL-3.0+
  */
 
-namespace ContaoCommunityAlliance\Composer\Plugin;
+namespace ContaoCommunityAlliance\Composer\Plugin\Exception;
 
 /**
- * Exception when composer try to install contao/core in vendor,
- * even if it's installed in parent directory.
+ * Exception when the configuration was updated through the pre-update scripts.
  */
-class DuplicateContaoException extends \RuntimeException
+class ConfigUpdateException extends \RuntimeException
 {
 }
